refactor(solver): extract result presentation into ResultPresenter
Move the render/applyRender/pendingRender/lastVirtualResult cluster out of
createSolverActions into createResultPresenter. The presenter owns how a solver
result becomes result-panel store state: virtual-vs-blocks shaping, widest-line
measurement, the rotation row cap, and the defer-while-navigating buffer, behind
a render/renderError/flushDeferred/clearResult/restoreFullVirtualResult API.

The two mutable locals (lastVirtualResult, pendingRender) and four closures
that shared them now live together instead of being threaded through
solverActions. The solver-error block construction folds into renderError.

Behavior-preserving (render logic relocated verbatim); tsc/test(73)/knip/build
green and bench within thresholds (p50 8.3ms, framesOver34:0).
